Flask_API_Server/requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
netifaces==0.11.0
numpy==2.2.2
Pillow==11.1.0
Requests==2.32.3
